Fix: Broken live render of interactive docs.
`datum.size` was undefined for non-leaf nodes of the tree.  I don't know
why this broke all of a sudden, I don't know why it continued to work
when server rendering, and I don't know how it ever worked.

I was only assigning a size of `1` to leaf nodes - this caused the arrangement
of arcs to match the design I wanted.  However, when calculating
datum.size was undefined for some nodes, and this value was used directly
to create the `node.value` value.

With an `undefined` value, d3-hierarchy's partition would fail and give
NaN for what should have been a number.  It would've happened here:

https://github.com/d3/d3-hierarchy/blob/e44f8a43c6e4bf89b974e0e799de03bcbb973165/src/treemap/dice.js#L6
https://github.com/d3/d3-hierarchy/blob/e44f8a43c6e4bf89b974e0e799de03bcbb973165/src/treemap/dice.js#L10

In any case, the solution was to explicitly provide a summable value for each
node: for leaf nodes, a value of 1; for non-leaf nodes, a value of 0;